Expect build.sh scripts to generate a target dir
When a `build.sh` script is present it is expected to generate a ./target directory and that target directory is expected to be used as the buildpack. This is present in the reference implementation https://github.com/heroku/buildpacks-jvm/blob/8ac7ba3a2607b0b6fcf20da15437470dc9a55512/test/rapier/rapier.rb#L39.